refactor: create util fun to count total nutrition
